before

+++ resources/views/cms/assets/list.blade.php
@@ -20,6 +20,14 @@ Assets
 
 <div class="row">
     <div class="col-sm col-lg">
+
+        @if ($message = Session::get('success'))
+        <div class="alert alert-success alert-block">
+            <button type="button" class="close" data-dismiss="alert">×</button>
+                <strong>{{ $message }}</strong>
+        </div>
+        @endif
+
         <table id="example" class="table table-striped table-bordered">
             <thead>
                 <tr>
@@ -60,14 +68,19 @@ Assets
     } );
 
     $(".btn-delete").click(function() {
-        $.ajax({
-            url: $(this).data('href'),
-            type: 'DELETE',
-            success: function(response) {
-                console.log(response);
-                location.reload();
-            }
-        });
+       
+        if (confirm('Are you sure you want to delete?')) {
+            
+            $.ajax({
+                url: $(this).data('href'),
+                type: 'DELETE',
+                success: function(response) {
+                    console.log(response);
+                    location.reload();
+                }
+            });
+        }
+       
     });
 </script>
